requirements.txt
django
mysqlclient
PyMySQL
djangorestframework
